controller: updated cotroller for approved or reject all types of request
