+++ controllers/requests.ts
@@ -1,30 +1,21 @@
 import {
   ERROR_WHILE_FETCHING_REQUEST,
-  ERROR_WHILE_UPDATING_REQUEST,
-  REQUEST_REJECTED_SUCCESSFULLY,
-  REQUEST_APPROVED_SUCCESSFULLY,
   REQUEST_FETCHED_SUCCESSFULLY,
-  REQUEST_STATE,
-  LOG_ACTION,
-  REQUEST_LOG_TYPE,
   REQUEST_TYPE,
 } from "../constants/requests";
-import { statusState } from "../constants/userStatus";
-import { addFutureStatus } from "../models/userStatus";
-import { createUserFutureStatus } from "../models/userFutureStatus";
-import { getRequests, updateRequest } from "../models/requests";
-import { addLog } from "../models/logs";
+import { getRequests } from "../models/requests";
 import { getPaginatedLink } from "../utils/helper";
-import { createOooRequestController } from "./oooRequests";
+import { createOooRequestController, updateOooRequestController } from "./oooRequests";
 import { OooRequestCreateRequest, OooRequestResponse } from "../types/oooRequest";
 import { CustomResponse } from "../typeDefinitions/global";
 import { ExtensionRequestRequest, ExtensionRequestResponse } from "../types/extensionRequests";
-import { createTaskExtensionRequest } from "./extensionRequestsv2";
+import { createTaskExtensionRequest, updateTaskExtensionRequest } from "./extensionRequestsv2";
+import { UpdateRequest } from "../types/requests";
 
 export const createRequestController = async (
   req: OooRequestCreateRequest | ExtensionRequestRequest,
-  res: CustomResponse) => {
-
+  res: CustomResponse
+) => {
   const type = req.body.type;
   switch (type) {
     case REQUEST_TYPE.OOO:
@@ -36,63 +27,15 @@ export const createRequestController = async (
   }
 };
 
-export const updateRequestController = async (req: any, res: any) => {
-  const requestBody = req.body;
-  const userId = req?.userData?.id;
-  const requestId = req.params.id;
-  if (!userId) {
-    return res.boom.unauthorized();
-  }
-
-  try {
-    const requestResult = await updateRequest(requestId, requestBody, userId);
-    if ("error" in requestResult) {
-      return res.boom.badRequest(requestResult.error);
-    }
-    const [logType, returnMessage] =
-      requestResult.state === REQUEST_STATE.APPROVED
-        ? [REQUEST_LOG_TYPE.REQUEST_APPROVED, REQUEST_APPROVED_SUCCESSFULLY]
-        : [REQUEST_LOG_TYPE.REQUEST_REJECTED, REQUEST_REJECTED_SUCCESSFULLY];
-
-    const requestLog = {
-      type: logType,
-      meta: {
-        requestId: requestId,
-        action: LOG_ACTION.UPDATE,
-        createdBy: userId,
-        createdAt: Date.now(),
-      },
-      body: requestResult,
-    };
-    await addLog(requestLog.type, requestLog.meta, requestLog.body);
-    if (requestResult.state === REQUEST_STATE.APPROVED) {
-      const requestData = await getRequests({ id: requestId });
-
-      if (requestData) {
-        const { from, until, requestedBy, message } = requestData as any;
-        const userFutureStatusData = {
-          requestId,
-          status: REQUEST_TYPE.OOO,
-          state: statusState.UPCOMING,
-          from,
-          endsOn: until,
-          userId: requestedBy,
-          message,
-        };
-        await createUserFutureStatus(userFutureStatusData);
-        await addFutureStatus(userFutureStatusData);
-      }
-    }
-    return res.status(201).json({
-      message: returnMessage,
-      data: {
-        id: requestResult.id,
-        ...requestResult,
-      },
-    });
-  } catch (err) {
-    logger.error(ERROR_WHILE_UPDATING_REQUEST, err);
-    return res.boom.badImplementation(ERROR_WHILE_UPDATING_REQUEST);
+export const updateRequestController = async (req: UpdateRequest, res: CustomResponse) => {
+  const type = req.body.type;
+  switch (type) {
+    case REQUEST_TYPE.OOO:
+      return await updateOooRequestController(req as UpdateRequest, res as ExtensionRequestResponse);
+    case REQUEST_TYPE.EXTENSION:
+      return await updateTaskExtensionRequest(req as UpdateRequest, res as ExtensionRequestResponse);
+    default:
+      return res.boom.badRequest("Invalid request type");
   }
 };
